pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]

[project]
authors = [{email = "ra@tcp.direct"}, {name = "ra"}]
classifiers = ["Development Status :: 0.4.1 Dev"]
dependencies = [
  "attrs>=21.4.0",
  "loguru>=0.6.0",
  "pycryptodomex>=3.14.1",
]
description = "MaPy MapleStory Emulator"
keywords = ["MapleStory", "Server Emulator"]
maintainers = [{name = "ra", email = "ra@tcp.direct"}]
name = "MaPy"
readme = "README.md"
requires-python = "^3.10.2"
version = "0.4.1"

[project.optional-dependencies]
linux = ["uvloop>=0.16.0"]

[tool.flake8]
exclude = "mapy/scripts/*/*.py"
extend-ignore = ["E203", "E501"]
max-line-length = 88

[tool.black]
include = "\\.pyi?$"
skip-magic-trailing-comma = true
target-version = ["py310"]
line-length = 88

[tool.poetry]
description = "MaPy MapleStory Emulator"
name = "MaPy"
version = "0.0.1"

license = "MIT"

authors = ["ra <ra@tcp.direct>"]

readme = "README.md"

homepage = "https://github.com/Rooba/mapy"
repository = "https://github.com/Rooba/mapy"

keywords = ["MapleStory", "Server Emulator"]

[tool.poetry.dependencies]
python = "^3.10" # Compatible python versions must be declared here
toml = "^0.10.2"
# Dependencies with extras
requests = {version = "^2.13", extras = ["security"]}
# Python specific dependencies with prereleases allowed
pathlib2 = {version = "^2.2", python = "~2.7", allow-prereleases = true}
# Git dependencies
cleo = {version = "~0.8.1"}

PyYAML = ">=6.0"
attrs = ">=21.4.0"
cryptography = ">=36.0.2"
loguru = ">=0.6.0"
pycryptodomex = ">=3.14.1"
pyright = ">=1.1.237"
yarl = ">=1.7.0"
uvloop = {version = ">=0.16.0", optional = true }
yapf = {version = ">=0.32.0", optional = true }
httptools = ">=0.4.0"
websockets = ">=10.3"
asyncpg = "^0.26.0"
sanic = "^22.6.2"

